Project 1/requirements.txt
cffi==1.14.4
pycparser==2.20
PyNaCl==1.4.0
six==1.15.0
